% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHCProcessing.R
\name{areaplot}
\alias{areaplot}
\title{plotting the area percentage}
\usage{
areaplot(..., type = c("box", "violin", "bar"))
}
\arguments{
\item{...}{the organized count data of each type of phenotype}

\item{type}{which kind of plot you want to output}
}
\value{
a boxplot/violin plot/barplot of mega area proprtion
}
\description{
This function is used organizing and plotting the area prportion: the area that megakaryocytes
occupied in all area occupied by cells
}
\details{
This function is used organizing and plotting the area prportion: the area that megakaryocytes
occupied in all area occupied by cells. This function will return a boxplot/violin plot/barplot of mega area proprtion
}
\author{
Ruitao Zhang
}
