% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHCProcessing.R
\docType{package}
\name{IHCProcessing}
\alias{IHCProcessing}
\title{IHCRProcessing: a series of functions for or ImageJ CD61 Megakaryocytes data processing.}
\description{
This packages is specifically developed for Linda Resar Lab to fully processing the CD61 IHC data or simliar
kind of cell recognition raw data output from ImageJ. You can input your .csv raw data, and use functions in
this package to reorganize, categorize, label and plot the data.
}
\details{
This package contains 5 functions:
\code{readraw} read the raw data and organize them into a list
\code{labind} bind and label each group of data
\code{numplot} plot number proportion
\code{areaplot} plot area proportion
\code{test} return p-value of multiple statistical test methoods
1. load the raw data into R
2. label and bind each groups and get the summary and percentage of mega and nuclei, respectively.
3. plot (which can be processed in R or Prism 8)
}
\section{Maintainer}{

Ruitao Zhang <rzhangh49@jhu.edu>
}

\author{
Ruitao Zhang
}
