% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHCProcessing.R
\name{numplot}
\alias{numplot}
\title{plotting the count percentage}
\usage{
numplot(..., type = c("box", "violin"))
}
\arguments{
\item{...}{the organized count data of each type of phenotype}

\item{type}{which kind of plot you want to output}
}
\value{
a boxplot/violin plot of mega number proprtion
}
\description{
This function is used organizing and plotting the number prportion: how many megakaryocytes in
all nuclei in this plot
}
\details{
This function is used organizing and plotting the number prportion: how many megakaryocytes in
all nuclei in this plot. This function will return a boxplot/violin plot of mega number proprtion
}
\author{
Ruitao Zhang
}
