% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHCProcessing.R
\name{readraw}
\alias{readraw}
\title{Read the Megakaryocytes and nuclei raw data from the document}
\usage{
readraw(wd)
}
\arguments{
\item{wd}{the working directory of the raw data}
}
\value{
a list with all groups of raw data
}
\description{
This function is used to read the raw data read from the document and make them a list
}
\details{
This function is used to read the raw data read from the document and make them a list. Each element of the list is a plot.
}
\author{
Ruitao Zhang
}
