% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHCProcessing.R
\name{test}
\alias{test}
\title{Get the p-value of multiple statistical test methoods}
\usage{
test(a, b, statistics = c("Student", "Mann", "Chisq", "Kolmogorov", "Fisher"))
}
\arguments{
\item{a}{the first data list you want to test}

\item{b}{the second data list you want to test}

\item{statistics}{the methods you use for statistical test}
}
\value{
the p-value of the test result
}
\description{
This function provides multiple statistical test methods for the unpaired-two sample testing.
}
\details{
This function provides multiple statistical test methods for the unpaired-two sample testing. And it will give
us the p-value of the test result as an output.
}
\author{
Ruitao Zhang
}
